Update: Dedup chartBuilders
